remove group configs from boot node RPC configs
